Calc residuals for 3d array: use generic linreg func & add test
