+++ R/RcppExports.R
@@ -1,10 +1,6 @@
 # This file was generated by Rcpp::compileAttributes
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-calc_resid_linreg_3d <- function(X, P) {
-    .Call('qtl2scan_calc_resid_linreg_3d', PACKAGE = 'qtl2scan', X, P)
-}
-
 interpolate_map <- function(oldpos, oldmap, newmap) {
     .Call('qtl2scan_interpolate_map', PACKAGE = 'qtl2scan', oldpos, oldmap, newmap)
 }
@@ -47,6 +43,10 @@ calc_rss_linreg <- function(X, Y) {
 
 calc_resid_linreg <- function(X, Y) {
     .Call('qtl2scan_calc_resid_linreg', PACKAGE = 'qtl2scan', X, Y)
+}
+
+calc_resid_linreg_3d <- function(X, P) {
+    .Call('qtl2scan_calc_resid_linreg_3d', PACKAGE = 'qtl2scan', X, P)
 }
 
 cbind_imatrix <- function(mat1, mat2) {
