--- src/hk_scan.h
@@ -1,4 +1,0 @@
-// simple genome scan by linear regression
-
-// use calc_resid_lapack for a 3-dim array
-NumericMatrix calc_resid_linreg_3d(NumericMatrix& X, NumericVector& P);
